refactor: Move DefaultDispatcherProvider and FirebaseDoorFcmRepository to shared modules
- Move DefaultDispatcherProvider to data/commonMain (pure coroutines, no Android deps)
- Move FirebaseDoorFcmRepository to data/commonMain (change AppSettings → AppSettingsRepository)
- Both are now available for iOS via KMP
